full_project/data_capture/requirements.txt
requests==2.26.0
mysql-connector-python==8.0.26
werkzeug==0.16.1